+++ src/main/java/com/kyfexuwu/m3we/editor/component/NonResizableComponent.java
@@ -1,20 +1,19 @@
 package com.kyfexuwu.m3we.editor.component;
 
 import com.kyfexuwu.m3we.editor.Block;
-import com.kyfexuwu.m3we.editor.Color;
 
 public abstract class NonResizableComponent extends Component{
-    private final float _width;
-    private final float _height;
-    public NonResizableComponent(Block parent, float width, float height) {
+    private final double _width;
+    private final double _height;
+    public NonResizableComponent(Block parent, double width, double height) {
         super(parent);
         this._width=width;
         this._height=height;
     }
 
     @Override
-    public float width(boolean isolated) { return this._width; }
+    public double width(boolean isolated) { return this._width; }
 
     @Override
-    public float height(boolean isolated) { return this._height; }
+    public double height(boolean isolated) { return this._height; }
 }
